move tests to time_test.go
